Update README.md
What is needed if the student wants to use East US?  Will the app work or will it break because the app is hard coded to use West Europe region?  Important not to assume the participants will always set the region to West Europe, as I do see it's hard coded in the views.